Road Accident Analysis using power bi
The purpose of this Power BI dashboard is to analyze road accident data and identify patterns and trends that can help improve road safety. Imported the data into Power BI and cleaned it to ensure it was accurate and ready for analysis. The data includes information about accidents, such as the number of vehicles involved, severity of the accident, and the location and time of the accident.